feature_generation/neo-rna/requirements.txt
####### PIP requirements #######
#
# Run `pip install -r requirements.txt`


## ----- Basic ------
#

# Jinja2 is a templating engine for Python.
# :link http://jinja2.pocoo.org/
Jinja2

# The YAML file parser
PyYAML

#
bs4

# Multi-worker wrapper
futures


## ----- Bio Info ------
#

# Numerical Python (NumPy) - a fast implementation of arrays and associated array functionality
# :link http://numpy.scipy.org/
numpy

# Biopython - tools for biological computation
# :link https://biopython.org/
biopython


## ----- Analytical Tools ------
#

# Simple and efficient tools for data mining and data analysis
# :link http://scikit-learn.org/stable/index.html
scikit-learn


# ReportLab - a library for generating PDF documents.
# :link http://www.reportlab.com/


## ----- Pipeline Tools ------
#

# pipeline framework
PyPPL


## ----- Others ------
#
# Python 2.6 compatibility
# argparse==1.2.1




